requirements.txt
https://bitbucket.org/kds_consulting_team/keboola-python-util-lib/get/0.5.1.zip#egg=kbc
mock
freezegun
requests
xmltodict
pandas
keboola.component
